+++ scheduler/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.43.2] - 2020-07-30
+### Fixed
+- Using 1024*1024 (mebibytes) as the k8s memory multiplier, from @dposada
+- Avoiding NPE due to missing resources when totaling resources for metrics, from @dposada 
+
 ## [1.43.1] - 2020-07-24
 ### Added
 - Per-pool global quotas, from @scrosby